Port src/marks/{density,difference} from JS to TS (stage 1)
Inline each .d.ts into its companion .js to produce a single .ts source per
mark, with no behavior changes. Internal helpers that the public .d.ts files
don't expose are imported with @ts-ignore so the runtime imports stay intact.
